requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
deepseek==1.0.0
easyocr==1.7.2
exceptiongroup==1.2.2
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
ollama==0.4.7
opencv-python-headless==4.11.0.86
packaging==24.2
pgvector==0.3.6
pillow==11.1.0
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
python-bidi==0.6.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.1
scipy==1.15.1
shapely==2.0.7
sniffio==1.3.1
sympy==1.13.1
tifffile==2025.1.10
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
